PYTHON DATA RESTRUCTURE AND CODE MANIPULATION
Dans ce commit, nous avons restructuré l'arborescence de notre application, en transformant toutes les dossiers en packages pour bien importer les modules que nous avons besoin. Aussi nous avons réussi à connecter à notre base des données à partir de notre application, mais nous n'avons pas encore réussi à lancer les tests.
