+++ .gitignore
@@ -1,2 +1,2 @@
 Application/python/dbconnection.py
-Application/python/.env
+Application/data/.env
