+++ Application/data/.gitignore
@@ -1,0 +1,3 @@
+# created by virtualenv automatically
+venv
+.env
