+++ Application/gui_app/__pycache__/app.cpython-39.pyc
