--- Application/python/__init__.py
@@ -1,2 +1,0 @@
-from .dbconnection import *
-from .main import *
